requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-cors-headers==4.2.0
django-debug-toolbar==4.1.0
django-ratelimit==4.1.0
django-redis==5.3.0
djangorestframework==3.14.0
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
psycopg2==2.9.7
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.4
yarl==1.9.2
